feat: extract noether-isolation crate + ship noether-sandbox binary
Delivers the v0.7.1 close-out promised to agentspec in noether#36.

## What moved

Everything under `noether_engine::executor::isolation` now lives in
a new top-level crate `noether-isolation`. The `noether-engine`
module becomes a thin `pub use` re-export so existing callers see
zero API change — this is a refactor for *other* consumers, not for
the noether-engine surface.

Public surface of `noether-isolation`:

- `IsolationBackend::{None, Bwrap{bwrap_path}}` + `auto()` / `from_flag()`
- `IsolationError::{UnknownBackend, BackendUnavailable}`
- `IsolationPolicy` (Serde-enabled — was `Debug+Clone` only)
- `IsolationPolicy::from_effects(&EffectSet)` / `.with_work_host(...)`
- `build_bwrap_command(bwrap, policy, inner_cmd) -> Command`
- `find_bwrap()` trusted-path-first discovery
- Constants: `NOBODY_UID`, `NOBODY_GID`, `TRUSTED_BWRAP_PATHS`

Dependency footprint is intentionally tiny: `noether-core` for
`Effect` / `EffectSet`, plus `serde`, `thiserror`, `tracing`. A
downstream consumer (agentspec, a future Python binding, a
standalone CLI) picks this up without pulling the composition
engine, semantic index, LLM providers, or Nix executor.

## New binary: noether-sandbox

~150 LOC, under `crates/noether-sandbox/`. Takes an
`IsolationPolicy` as JSON on stdin plus an argv after `--`, runs
the argv inside the bwrap sandbox using
`noether_isolation::build_bwrap_command`, propagates the child's
exit code.

```bash
echo '{"ro_binds":[["/nix/store","/nix/store"]],"network":true,"env_allowlist":["PATH"]}' \
  | noether-sandbox -- claude-code -p "hello"
```

- `--isolate=auto|bwrap|none` / `NOETHER_ISOLATION` env var
  mirrors `noether run`.
- Empty stdin falls back to `IsolationPolicy::from_effects(&pure)`
  — sensible default for the "sandbox me an echo" case.
- stdin/stdout/stderr pass through to the inner command
  (interactive tty works).

End-to-end smoke test on this host confirms
`os.getuid() == 65534` inside the sandbox as expected. All 12
tests pass (isolation unit tests + sandbox argv-parse tests).

## Changes to IsolationPolicy

- Added `#[derive(Serialize, Deserialize)]` with
  `#[serde(default, skip_serializing_if = "Option::is_none")]` on
  `work_host` so the wire format stays compact. Round-trip test
  pins the shape.
- `NOBODY_UID`, `NOBODY_GID`, `TRUSTED_BWRAP_PATHS` promoted from
  `pub(crate)` to `pub` — downstream consumers may want to
  introspect the conventions without re-implementing them.

## Workspace

- New members: `crates/noether-isolation`, `crates/noether-sandbox`.
- `noether-engine` now depends on `noether-isolation` transitively.
- Workspace version bumped 0.7.0 → 0.7.1.
- CHANGELOG.md gains a 0.7.1 section.

## Tests

- 11 isolation tests (unchanged, migrated with the module).
- 1 new `policy_round_trips_through_json` test pinning the serde
  shape — the wire format is the API contract for non-Rust
  consumers, so a future field reorder or rename is a breaking
  change and this test makes it visible.
- 6 new sandbox-binary argv-parse tests.
- Workspace: `cargo test --workspace` green, `cargo clippy
  --all-targets -- -D warnings` clean.

## Follow-up

- Update agentspec issue #36 with the v0.7.1 landing + point at the
  extracted crate.
- noether-cloud#26 can now optionally depend on `noether-isolation`
  directly instead of `noether-engine` — not critical since the
  registry doesn't run NixExecutor today, but cleanly narrows the
  dependency footprint if/when it does.
