+++ crates/noether-engine/Cargo.toml
@@ -19,6 +19,7 @@ native = ["dep:reqwest", "dep:rusqlite", "dep:arrow", "dep:base64"]
 [dependencies]
 noether-core = { path = "../noether-core", version = "0.7" }
 noether-store = { path = "../noether-store", version = "0.7" }
+noether-isolation = { path = "../noether-isolation", version = "0.7" }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_jcs = "0.2"
